+++ PlantVsZombie/src/dut/game/GameObject.java
@@ -14,6 +14,6 @@ public interface GameObject {
 	public boolean isAlive();
 	public Color getColor();
 	public boolean matrixOut(GameView v);
-	public void addToHealth(int i);
-	public int getDamage();
+	public void addToHealth(double i);
+	public double getDamage();
 }
